removed licenceFactory moved logic to condition service. moved conditions logic out of licence service. Added preliminary new conditions config
